refactor packages, create unit tests, create e2e route test
