+++ hnclient-v2/pom.xml
@@ -100,14 +100,19 @@
       <artifactId>camel-test</artifactId>
       <scope>test</scope>
     </dependency>
-    
     <dependency>
         <groupId>org.junit.jupiter</groupId>
         <artifactId>junit-jupiter-engine</artifactId>
         <version>5.5.2</version>
         <scope>test</scope>
     </dependency>
-    
+    <dependency>
+        <groupId>org.mockito</groupId>
+        <artifactId>mockito-core</artifactId>
+        <version>3.6.28</version>
+        <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
   <build>
