+++ hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/authentication/ClientAuthenticationBuilder.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnclientv2.auth;
+package ca.bc.gov.hlth.hnclientv2.authentication;
 
 import com.nimbusds.oauth2.sdk.auth.ClientAuthentication;
