+++ hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/authentication/RetrieveAccessToken.java
@@ -1,8 +1,7 @@
-package ca.bc.gov.hlth.hnclientv2;
+package ca.bc.gov.hlth.hnclientv2.authentication;
 
-import ca.bc.gov.hlth.error.CamelCustomException;
-import ca.bc.gov.hlth.error.ServerNoConnectionException;
-import ca.bc.gov.hlth.hnclientv2.auth.ClientAuthenticationBuilder;
+import ca.bc.gov.hlth.hnclientv2.error.CamelCustomException;
+import ca.bc.gov.hlth.hnclientv2.error.ServerNoConnectionException;
 import com.nimbusds.oauth2.sdk.*;
 import com.nimbusds.oauth2.sdk.auth.ClientAuthentication;
 import com.nimbusds.oauth2.sdk.token.AccessToken;
