+++ hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/error/ErrorCodes.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.error;
+package ca.bc.gov.hlth.hnclientv2.error;
 
 import java.util.HashMap;
 import java.util.Map;
