+++ hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/error/MessageUtil.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnclientv2;
+package ca.bc.gov.hlth.hnclientv2.error;
 
 public class MessageUtil {
