+++ hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/handshakeserver/HandshakeServer.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnclientv2.handshake.server;
+package ca.bc.gov.hlth.hnclientv2.handshakeserver;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -14,12 +14,14 @@ import org.apache.camel.ProducerTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ca.bc.gov.hlth.error.ErrorBuilder;
-import ca.bc.gov.hlth.hnclientv2.MessageUtil;
+import ca.bc.gov.hlth.hnclientv2.error.ErrorBuilder;
+import ca.bc.gov.hlth.hnclientv2.error.MessageUtil;
 import ca.bc.gov.hlth.hnclientv2.Util;
 import io.netty.util.internal.StringUtil;
 
 public class HandshakeServer {
+
+	private static final Logger logger = LoggerFactory.getLogger(HandshakeServer.class);
 
 	private static final int XFER_HANDSHAKE_SEED = 0;
 	private static final int XFER_HANDSHAKE_SIZE = 8;
@@ -35,22 +37,19 @@ public class HandshakeServer {
 
 	private byte decodeSeed = 0;
 	// DT segments to send to POS
-	private byte[] dataSegmentout = new byte[12];
+	private byte[] dataSegmentOut = new byte[12];
 	private byte[] dataHL7out;
 
 	private static final String TEN_ZEROS = "0000000000";
 
-	private static final Logger logger = LoggerFactory.getLogger(HandshakeServer.class);
-
 	private final ProducerTemplate producer;
 
-	private static final Util util = new Util();
+	private static final HandshakeUtil util = new HandshakeUtil();
 
 	public HandshakeServer(ProducerTemplate producer) {
 		logger.debug("HandshakeServer constructor is called");
 		this.producer = producer;
 		startServer();
-
 	}
 
 	public void startServer() {
@@ -178,14 +177,14 @@ public class HandshakeServer {
 							}
 							String dtSegment = insertHeader(hnSecureResponse);
 
-							dataSegmentout = dtSegment.substring(0, 12).getBytes(StandardCharsets.UTF_8);
+							dataSegmentOut = dtSegment.substring(0, 12).getBytes(StandardCharsets.UTF_8);
 
 							dataHL7out = dtSegment.substring(12).getBytes(StandardCharsets.UTF_8);
 							logger.info("dataHL7out: " + new String(dataHL7out));
 
 							// scramble each segment prior to sending it to BufferedOutputStream
-							util.scrambleData(dataSegmentout, decodeSeed);
-							socketOutput.write(dataSegmentout);
+							util.scrambleData(dataSegmentOut, decodeSeed);
+							socketOutput.write(dataSegmentOut);
 
 							util.scrambleData(dataHL7out, decodeSeed);
 							socketOutput.write(dataHL7out);
