+++ hnclient-v2/src/main/resources/log4j2.properties
@@ -9,5 +9,5 @@ rootLogger.appenderRef.out.ref = out
 logger.app.name = HttpLogger
 logger.app.level = DEBUG
 
-#logger.retrieveaccesstoken.name = ca.bc.gov.hlth.hnclientv2.RetrieveAccessToken
+#logger.retrieveaccesstoken.name = ca.bc.gov.hlth.hnclientv2.authentication.RetrieveAccessToken
 #logger.retrieveaccesstoken.level = DEBUG
