+++ hnclient-v2/src/test/java/ca/bc/gov/hlth/hnclientv2/wrapper/FHIRJsonUtilTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package ca.bc.gov.hlth.hnclientv2.unit;
+package ca.bc.gov.hlth.hnclientv2.wrapper;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -12,8 +12,8 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.Test;
 
-import ca.bc.gov.hlth.hnclientv2.json.FHIRJsonMessage;
-import ca.bc.gov.hlth.hnclientv2.json.FHIRJsonUtil;
+import ca.bc.gov.hlth.hnclientv2.wrapper.FHIRJsonMessage;
+import ca.bc.gov.hlth.hnclientv2.wrapper.FHIRJsonUtil;
 import net.minidev.json.JSONObject;
 import net.minidev.json.JSONValue;
 
@@ -23,8 +23,6 @@ import net.minidev.json.JSONValue;
  *
  */
 public class FHIRJsonUtilTest {
-	
-	public String hl7NullValue = null;
 	
 	public String msgE45 = "MSH|^~\\&|HR|BC00000098|RAIGET-DOC-SUM|BC0003000|19991004103039|lharris|E45|19980915000015|D|2.3\r\n" + 
 			"HDR|||TRAININGAdmin\r\n" + 
@@ -37,42 +35,36 @@ public class FHIRJsonUtilTest {
 			"PID||9081361205^^^BC^PH\r\n" + 
 			"IN1||||||||3131182||123|^^1234|20010601\r\n" + 
 			"ZIA||||||||||||||||123 FIRST ST^TOWN BC^^^^^^^^^^^^^^^^^^^^V9A1H9^^H|^PH^PH^^^123^1235555~^PH^PH^^^999^1115555\r\n" + 
-			"NK1|||SP||||||||||||||||||||||||||||||9073094356^^^BC^PH\r\n" + 
-			"NK1|||DP||||||||||||||||||||||||||||||9876913098^^^BC^PH\r\n" + 
-			"NK1|||DP||||||||||||||||||||||||||||||9083410453^^^BC^PH";
+			"NK1|||SP||||||||||||||||||||||||||||||1233094356^^^BC^PH\r\n" +
+			"NK1|||DP||||||||||||||||||||||||||||||1236913098^^^BC^PH\r\n" +
+			"NK1|||DP||||||||||||||||||||||||||||||1233410453^^^BC^PH";
 	
 	public String msgR31="MSH|^~\\&|HR|BC00000098|RAIEST_PYR-RLDP|BC0003000|19991013153348|PM33283|R31|19980915000015|D|2.3\r\n" + 
 			"ZHD|19991013153348|^^00000001|TRAININGAdmin\r\n" + 
-			"PID||9074790865^^^BC345^PH|12345678901234567890^^^BC345^MH\r\n" + 
+			"PID||1234790865^^^BC345^PH|12345678901234567890^^^BC345^MH\r\n" +
 			"IN1||||||||1250000||||20010101\r\n" + 
 			"NK1|||12||||||||||||||||||||||||||||||1234567890^^^12345^12\r\n" + 
 			"ZSG|||||";
 	
 	@Test
 	public void testCreateFHIRJsonNullObject() {		
-		JSONObject jsonObj = FHIRJsonUtil.createFHIRJsonObj(hl7NullValue);
+		JSONObject jsonObj = FHIRJsonUtil.createFHIRJsonObj(null);
 		assertNull(jsonObj);
 	}
-	
+
 	@Test
-	public void testCreateFHIRJsonNotNullObject() {		
-		JSONObject jsonObj = FHIRJsonUtil.createFHIRJsonObj(msgE45);
-		assertNotNull(jsonObj);
+	public void testCreateFHIRJsonEmptyObject() {
+		JSONObject jsonObj = FHIRJsonUtil.createFHIRJsonObj("");
+		assertNull(jsonObj);
+	}
+
+	@Test
+	public void testDoNotThrowCreateFHIRJsonObj_msgE45()  {
+		assertDoesNotThrow(() -> FHIRJsonUtil.createFHIRJsonObj(msgE45));
 	}
 	
 	@Test
-	public void testShouldNotEmptyFHIRJsonObj_msgE45Obj()  {	
-		JSONObject fhirObj = FHIRJsonUtil.createFHIRJsonObj(msgE45);
-		assertFalse(fhirObj.isEmpty());
-	}
-	
-	@Test
-	public void testDoNotThrowCreateFHIRJsonObj_msgE45()  {	
-		assertDoesNotThrow(() -> FHIRJsonUtil.createFHIRJsonObj(msgE45));		
-	}
-	
-	@Test
-	public void testIsJSONMsgE45()  {	
+	public void testIsJSONMsg()  {
 		JSONObject fhirObj = FHIRJsonUtil.createFHIRJsonObj(msgE45);
 		assertTrue(JSONValue.isValidJson(fhirObj.toJSONString()));
 	}
@@ -89,20 +81,6 @@ public class FHIRJsonUtilTest {
 	public void testCreateFHIRJsonObj_msgr30() {
 		JSONObject fhirObj = FHIRJsonUtil.createFHIRJsonObj(msgr30);
 		String expectedJsonString = "{\"content\":[{\"attachment\":{\"data\":\"MSH|^~\\\\&|HR|BC00000098|RAIEST-PYR-RL|BC0003000|19990719125635|PM33283|R30|19980915000017|D|2.3\\r\\nZHD|19990719125635|^^00000001|TRAININGAdmin\\r\\nPID||9081361205^^^BC^PH\\r\\nIN1||||||||3131182||123|^^1234|20010601\\r\\nZIA||||||||||||||||123 FIRST ST^TOWN BC^^^^^^^^^^^^^^^^^^^^V9A1H9^^H|^PH^PH^^^123^1235555~^PH^PH^^^999^1115555\\r\\nNK1|||SP||||||||||||||||||||||||||||||9073094356^^^BC^PH\\r\\nNK1|||DP||||||||||||||||||||||||||||||9876913098^^^BC^PH\\r\\nNK1|||DP||||||||||||||||||||||||||||||9083410453^^^BC^PH\",\"contentType\":\"x-application\\/hl7-v2+er7\"}}],\"resourceType\":\"DocumentReference\",\"status\":\"current\"}";
-		String actualString = fhirObj.toJSONString();
-		assertEquals(expectedJsonString,actualString);
-	}
-
-	@Test
-	public void testIsJSONMsgr30()  {	
-		JSONObject fhirObj = FHIRJsonUtil.createFHIRJsonObj(msgr30);
-		assertTrue(JSONValue.isValidJson(fhirObj.toJSONString()));
-	}
-	
-	@Test
-	public void testCreateFHIRJsonObj_msgR31()  {
-		JSONObject fhirObj = FHIRJsonUtil.createFHIRJsonObj(msgR31);
-		String expectedJsonString = "{\"content\":[{\"attachment\":{\"data\":\"MSH|^~\\\\&|HR|BC00000098|RAIEST_PYR-RLDP|BC0003000|19991013153348|PM33283|R31|19980915000015|D|2.3\\r\\nZHD|19991013153348|^^00000001|TRAININGAdmin\\r\\nPID||9074790865^^^BC345^PH|12345678901234567890^^^BC345^MH\\r\\nIN1||||||||1250000||||20010101\\r\\nNK1|||12||||||||||||||||||||||||||||||1234567890^^^12345^12\\r\\nZSG|||||\",\"contentType\":\"x-application\\/hl7-v2+er7\"}}],\"resourceType\":\"DocumentReference\",\"status\":\"current\"}";
 		String actualString = fhirObj.toJSONString();
 		assertEquals(expectedJsonString,actualString);
 	}
