+++ hnclient-v2/src/test/java/ca/bc/gov/hlth/hnclientv2/wrapper/ProcessV2ToJsonTest.java
@@ -1,55 +1,54 @@
-/**
- * 
- */
-package ca.bc.gov.hlth.hnclientv2.route;
+package ca.bc.gov.hlth.hnclientv2.wrapper;
 
+import org.apache.camel.CamelExecutionException;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Test;
 
-import ca.bc.gov.hlth.hnclientv2.wrapper.ProcessV2ToJson;
-
-/**
- * @author Tony.Ma * 
- * @date Jan. 20, 2021
- *
- */
-public class ProcessV2ToJsonRouteTest extends CamelTestSupport {
-	
-	String TestV2Msg = "00000352MSH|^~\\&|HNWEB|VIHA|RAIGT-PRSN-DMGR|BC00001013|20170125122125|train96|R03|20170125122125|D|2.4||\n"
-			+ "ZHD|20170125122125|^^00000010|HNAIADMINISTRATION||||2.4\n"
-			+ "PID||1234567890^^^BC^PH";
+public class ProcessV2ToJsonTest extends CamelTestSupport {
 	
 	String encodeV2Msg = "MDAwMDAzNTJNU0h8Xn5cJnxITldFQnxWSUhBfFJBSUdULVBSU04tRE1HUnxCQzAwMDAxMDEzfDIwMTcwMTI1MTIyMTI1fHRyYWluOTZ8UjAzfDIwMTcwMTI1MTIyMTI1fER8Mi40fHwKWkhEfDIwMTcwMTI1MTIyMTI1fF5eMDAwMDAwMTB8SE5BSUFETUlOSVNUUkFUSU9OfHx8fDIuNApQSUR8fDEyMzQ1Njc4OTBeXl5CQ15QSA==";
-	
-	
+
+	// Define a route with the ProcessV2ToJson
 	@Override
-	protected RouteBuilder createRouteBuilder() throws Exception{
-
+	protected RouteBuilder createRouteBuilder() {
 		return new RouteBuilder() {
-
 			@Override
-			public void configure() throws Exception {			
+			public void configure() {
 				from("direct:sampleInput")
 				.process(new ProcessV2ToJson())
-                .log("v2Message decode")
-                .log("Sending to HNClient")
                 .to("mock:output");
-
 			}
 		};
 	}
-	
-	
+
 	@Test
 	public void testV2MsgFlow() throws Exception {
 		String expectedMsg = "{\"content\":[{\"attachment\":{\"data\":\"MDAwMDAzNTJNU0h8Xn5cJnxITldFQnxWSUhBfFJBSUdULVBSU04tRE1HUnxCQzAwMDAxMDEzfDIwMTcwMTI1MTIyMTI1fHRyYWluOTZ8UjAzfDIwMTcwMTI1MTIyMTI1fER8Mi40fHwKWkhEfDIwMTcwMTI1MTIyMTI1fF5eMDAwMDAwMTB8SE5BSUFETUlOSVNUUkFUSU9OfHx8fDIuNApQSUR8fDEyMzQ1Njc4OTBeXl5CQ15QSA==\",\"contentType\":\"x-application\\/hl7-v2+er7\"}}],\"resourceType\":\"DocumentReference\",\"status\":\"current\"}";
+		// Define the route endpoint
 		MockEndpoint mock = getMockEndpoint("mock:output");
+
+		// Expect the json encoded message at the endpoint
 		mock.expectedMessageCount(1);
-		mock.expectedBodiesReceived(expectedMsg);		
+		mock.expectedBodiesReceived(expectedMsg);
+
+		// When we send in a Base64Encoded v2 message
 	    template.sendBody("direct:sampleInput",encodeV2Msg );	
 	    assertMockEndpointsSatisfied();
+	}
+
+	@Test
+	public void processV2ToJson_nullMessageTest() {
+		try {
+			// When we send in a null v2 message to a route with the ProcessV2ToJson processor
+			template.sendBody("direct:sampleInput", null);
+		} catch (CamelExecutionException e ){
+			// We expect a null pointer exception
+			// Because we have no exception handling in our simple test route camel throws
+			// the IllegalArgumentException upwards as a CamelExecutionException so we check the cause
+			assertTrue(e.getCause() instanceof IllegalArgumentException);
+		}
 	}
